backend/requirements.txt
asgiref~=3.5.2
django~=4.1.4
django-cors-headers~=3.13.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
pyjwt~=2.6.0
pytz~=2022.6
sqlparse~=0.4.3
tzdata~=2022.7
pillow~=9.3.0